+++ Changelog.md
@@ -18,6 +18,7 @@ v2.32.0 (not published yet)
 - UI: Add HDR metadata extraction to project options
 - UI: Add HDR metadata file to encoder settings when using "Import VUI metadata"
 - UI: Search for HDR metadata file next to source file and in temp folder to let it import via "Import VUI metadata"
+- UI: Add option to show an error message on "Frame Mismatch" after encoding
 - UI: Add "DefaultOrFirst" to "Default Subtitles" option
 - UI: Add "Close Project" to Main Menu
 - UI: Change default option for "Default Subtitles" to "Default"